added heavywater miketheman

+++ Rakefile
@@ -4,6 +4,7 @@ Bundler.require
 require 'yaml'
 require 'base64'
 require 'rake/clean'
+require 'pry'
 
 def cookbook_list(manifest='upstream/opscode_-_cookbooks.yml', scope='all')
   $stdout.puts "Starting build cookbook tasks: #{manifest}"
@@ -13,37 +14,40 @@ def cookbook_list(manifest='upstream/opscode_-_cookbooks.yml', scope='all')
     if File.directory?(cookbook_path)
       $stdout.puts "  Build Rake task: tmp/#{cookbook}"
 
-      file "tmp/#{cookbook}" do
-        $stdout.puts "  Cloning #{cookbook}"
-        git "clone --no-hardlinks cookbooks tmp/#{cookbook}"
-        Dir.chdir("#{Rake.original_dir}/tmp/#{cookbook}") do
-          puts `pwd`
-          $stdout.puts "  Extracting #{cookbook}"
-          git "filter-branch --subdirectory-filter #{cookbook} HEAD -- --all"
-          git "reset --hard"
-          git 'reflog expire --expire=now --all'
-          git 'repack -ad'
-          git "gc --aggressive --prune=now"
+      unless File.directory?("#{Rake.original_dir}/tmp/#{cookbook}")
 
-          create_repo(cookbook) unless repo_exists?(cookbook)
+        file "tmp/#{cookbook}" do
+          $stdout.puts "  Cloning #{cookbook}"
+          git "clone --no-hardlinks cookbooks tmp/#{cookbook}"
+          Dir.chdir("#{Rake.original_dir}/tmp/#{cookbook}") do
+            puts `pwd`
+            $stdout.puts "  Extracting #{cookbook}"
+            git "filter-branch --subdirectory-filter #{cookbook} HEAD -- --all"
+            git "reset --hard"
+            git 'reflog expire --expire=now --all'
+            git 'repack -ad'
+            git "gc --aggressive --prune=now"
 
-          # check for existing tags
-          git "remote rm origin"
-          git "remote add cookbooks git@github.com:#{config['org']}/#{cookbook}.git"
-          git "fetch cookbooks"
+            create_repo(cookbook) unless repo_exists?(cookbook)
 
-          # tag versions
-          revisions = git_output "rev-list --topo-order --branches"
-          version = nil
-          revisions.split(/\n/).each do |rev|
-            metadata = parse_metadata(cookbook, rev)
-            if metadata['version'] && metadata['version'] != version
-              version = metadata['version']
-              puts "tagging #{rev} as #{version}"
-              git "tag -a #{version}  -m 'Chef cookbook #{cookbook} version: #{version}' #{rev}"
+            # check for existing tags
+            git "remote rm origin"
+            git "remote add cookbooks git@github.com:#{config['org']}/#{cookbook}.git"
+            git "fetch cookbooks"
+
+            # tag versions
+            revisions = git_output "rev-list --topo-order --branches"
+            version = nil
+            revisions.split(/\n/).each do |rev|
+              metadata = parse_metadata(cookbook, rev)
+              if metadata['version'] && metadata['version'] != version
+                version = metadata['version']
+                puts "tagging #{rev} as #{version}"
+                git "tag -a #{version}  -m 'Chef cookbook #{cookbook} version: #{version}' #{rev}"
+              end
             end
+            git "push --tags cookbooks master"
           end
-          git "push --tags cookbooks master"
         end
 
       end
@@ -77,6 +81,9 @@ def post(uri, payload)
   sleep 1 # github api throttlin'
   basic_auth = Base64.encode64("#{config['login']}/token:#{config['token']}").gsub("\n", '')
   headers = { 'Authorization' => "Basic #{basic_auth}", :content_type => :json, :accept => :json}
+  puts "URI: #{uri}"
+  puts "JSON payload: #{payload.to_json}"
+  puts "Headers: #{headers}"
   JSON.parse(RestClient.post(uri, payload.to_json, headers))
 end
 
@@ -107,13 +114,28 @@ def create_repo(name)
     :description => "A Chef cookbook for #{name} (Initial Upstream: #{@upstream}, Repository: #{@repository})",
     :homepage    => "https://github.com/opscode/cookbooks/blob/master/LICENSE"
   }
-  post "https://github.com/api/v2/json/repos/create", repo_info
-  post "https://github.com/api/v2/json/teams/#{config['team_id']}/repositories", {:name => "#{config['org']}/#{name}"}
+  create_repo_uri = "https://github.com/api/v2/json/repos/create"
+  create_repo_result = post create_repo_uri, repo_info
+  puts create_repo_result.inspect
+  create_team_uri = "https://github.com/api/v2/json/organizations/#{config['org']}/teams/"
+  team_info = {:team => { :name => name, :permission => 'push', :repo_names => ["#{config['org']}/#{name}"] } }
+  sleep 5
+  create_team_result = post create_team_uri, team_info
+  # team_id = create_team_result['team']['id']
+  # add_team_member_uri = "https://github.com/api/v2/json/teams/#{team_id}/members?name=#{config['login']}"
+  # add_team_member_result = post add_team_member_uri
+  # puts add_team_member_result.inspect
+  create_team_result
 end
 
 def repo_exists?(name)
-  repositories = get("https://github.com/api/v2/json/organizations/repositories")
-  repositories['repositories'].detect { |r| r["name"] == name }
+  repo = true
+  begin
+    repositories = get("https://github.com/api/v2/json/repos/show/#{config['org']}/#{name}")
+  rescue RestClient::Exception => e
+    repo = false if e.http_body[/Repository not found/]
+  end
+  repo
 end
 
 def parse_metadata(cookbook, rev)
@@ -168,6 +190,10 @@ task :default do |tsk|
   begin
     update_all
   ensure
+    CLEAN.include('tmp/*')
+    Rake::Task['clone_clean'].invoke
+    Rake::Task['clean'].invoke
+    cleanup_clone()
   end
 end
 
@@ -183,7 +209,7 @@ task :create_tasks, [:cookbook] do |tsk, args|
   end
 end
 
-desc "Update specific cookbook (default: all) from Opscode's Chef Cookbooks repository."
+desc "Update cookbooks (default: all) from <upstream> Chef Cookbooks repositories."
 task :update, [:cookbook] => [:clone_clean] do |tsk, args|
   begin
     args.with_defaults(:cookbook => 'all')
@@ -201,4 +227,9 @@ task :update, [:cookbook] => [:clone_clean] do |tsk, args|
   end
 end
 
-cookbook_list
+#Dir.chdir(Rake.original_dir) do |path|
+#  FileList["upstream/*.yml"].collect do |manifest|
+#    $stdout.puts "Processing Cookbook manifest: #{manifest} "
+#    cookbook_list(manifest, args[:cookbook])
+#  end
+#end
